Move payment methods folder to purchases
