+++ client/me/purchases/payment-methods/controller.js
@@ -3,7 +3,7 @@
  */
 
 import React from 'react';
-import PaymentMethodsComponent from './main';
+import PaymentMethodsComponent from 'calypso/me/purchases/payment-methods/main';
 
 export function paymentMethods( context, next ) {
 	context.primary = React.createElement( PaymentMethodsComponent );
